Test: article e2e에 있던 get-commnet api 테스트 이전
- 댓글 가져오는 api를 comment 모듈로 이전하면서 테스트도 그에 맞게 수정
