+++ apps/api/test/e2e/comment.e2e-spec.ts
@@ -4,8 +4,10 @@ import { AuthModule } from '@api/auth/auth.module';
 import { AuthService } from '@api/auth/auth.service';
 import { CategoryModule } from '@api/category/category.module';
 import { CategoryRepository } from '@api/category/repositories/category.repository';
-import { CommentModule } from '@api/comment/comment.module';
+import { CommentApiModule } from '@api/comment/comment-api.module';
+import { CommentRepository } from '@api/comment/repositories/comment.repository';
 import { UserRepository } from '@api/user/repositories/user.repository';
+import { ANONY_USER_CHARACTER, ANONY_USER_ID, ANONY_USER_NICKNAME } from '@api/user/user.constant';
 import { UserModule } from '@api/user/user.module';
 import { Article } from '@app/entity/article/article.entity';
 import { Comment } from '@app/entity/comment/comment.entity';
@@ -14,11 +16,10 @@ import { HttpStatus, INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { E2eTestBaseModule } from '@test/e2e/e2e-test.base.module';
 import { clearDB, createTestApp } from '@test/e2e/utils/utils';
+import { testDto } from '@test/e2e/utils/validate-test';
 import * as request from 'supertest';
 import { getConnection } from 'typeorm';
-import { CommentRepository } from './../../src/comment/repositories/comment.repository';
 import * as dummy from './utils/dummy';
-import {CommentApiModule} from "@api/comment/comment-api.module";
 
 describe('Comments', () => {
   let httpServer: INestApplication;
@@ -194,6 +195,110 @@ describe('Comments', () => {
         .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
 
       expect(response.status).toEqual(HttpStatus.NOT_FOUND);
+    });
+  });
+
+  describe('/articles/{id}/comments', () => {
+    let comments: dummy.DummyComments;
+
+    beforeEach(async () => {
+      comments = await dummy.createDummyComments(commentRepository, users, articles);
+      JWT = dummy.jwt(users.cadet[0], authService);
+    });
+
+    test('[성공] GET - 게시글 댓글 목록 조회', async () => {
+      const articleId = articles.first.id;
+
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+      expect(response.status).toBe(HttpStatus.OK);
+
+      const responseComments = response.body.data as Comment[];
+      expect(responseComments.length).toBe(2);
+      expect(responseComments[0].id).toBe(comments.second.id);
+      expect(responseComments[0].content).toBe(comments.second.content);
+      expect(responseComments[0].writerId).toBe(comments.second.writerId);
+      expect(responseComments[0].writer.id).toBe(comments.second.writerId);
+      expect(responseComments[0].writer.nickname).toBe(users.cadet[1].nickname);
+      expect(responseComments[0].articleId).toBe(comments.second.articleId);
+      expect(responseComments[0].likeCount).toBe(comments.second.likeCount);
+      expect(responseComments[1].id).toBe(comments.first.id);
+    });
+
+    // TODO: 여기 테스트 기능 구현할것
+    test.skip('[성공] GET - 익명 게시글 댓글 목록 조희', async () => {
+      const articleId = articles.second.id;
+
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+      expect(response.status).toBe(HttpStatus.OK);
+
+      const responseComments = response.body.data as Comment[];
+      expect(responseComments.length).toBe(2);
+      expect(responseComments[0].id).toBe(comments[1].id);
+      expect(responseComments[0].content).toBe(comments[1].content);
+      expect(responseComments[0].writerId).toBe(ANONY_USER_ID);
+      expect(responseComments[0].writer.id).toBe(ANONY_USER_ID);
+      expect(responseComments[0].writer.nickname).toBe(ANONY_USER_NICKNAME);
+      expect(responseComments[0].writer.character).toBe(ANONY_USER_CHARACTER);
+      expect(responseComments[0].articleId).toBe(comments[1].articleId);
+      expect(responseComments[0].likeCount).toBe(comments[1].likeCount);
+      expect(responseComments[1].id).toBe(comments[0].id);
+    });
+
+    test('[성공] GET - GUEST 댓글 목록 조회', async () => {
+      const articleId = articles.first.id;
+
+      const response = await request(httpServer).get(`/articles/${articleId}/comments`);
+
+      expect(response.status).toBe(HttpStatus.FORBIDDEN);
+    });
+
+    test('[성공] GET - 게시글 댓글 목록 조회 권한 높은사람', async () => {
+      const articleId = articles.first.id;
+
+      JWT = dummy.jwt(users.admin[0], authService);
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+      expect(response.status).toBe(HttpStatus.OK);
+    });
+
+    test('[실패] GET - 게시글 댓글 목록 조회 권한 낮은사람', async () => {
+      const articleId = articles.first.id;
+
+      JWT = dummy.jwt(users.novice[0], authService);
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+      expect(response.status).toBe(HttpStatus.FORBIDDEN);
+    });
+
+    test('[실패] GET - 게시글 댓글 목록 존재하지 않는 게시글', async () => {
+      const articleId = 99;
+
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+
+      expect(response.status).toBe(HttpStatus.NOT_FOUND);
+    });
+
+    testDto<{ articleId: number }>([
+      ['articleId', undefined],
+      ['articleId', 'abc'],
+    ])('[실패] GET - 게시글 댓글 목록 %s인 경우', async (_, buildDto) => {
+      const articleId = buildDto({
+        articleId: articles.first.id,
+      }).articleId;
+
+      const response = await request(httpServer)
+        .get(`/articles/${articleId}/comments`)
+        .set('Cookie', `${process.env.ACCESS_TOKEN_KEY}=${JWT}`);
+
+      expect(response.status).toBe(HttpStatus.BAD_REQUEST);
     });
   });
 });
