+++ apps/api/test/e2e/reaction.e2e-spec.ts
@@ -37,7 +37,6 @@ describe('Reaction', () => {
         AuthModule,
         ArticleModule,
         CategoryModule,
-        CommentApiModule,
         ReactionModule,
       ],
     }).compile();
